requirements.txt
inferactively-pymdp==0.0.7.1
rich==13.7.0